committ

+++ app/Http/Controllers/Admin/AdminPlaceController.php
@@ -4,24 +4,13 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\Place;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
-class CategoryController extends Controller
+class AdminPlaceController extends Controller
 {
-    protected $appends = [
-        'getParentsTree'
-    ];
-    public static function getParentsTree($category, $title)
-    {
-        if ($category->parent_id == 0)
-        {
-            return $title;
-        }
-        $parent=Category::find($category->parent_id);
-        $title=$parent->title. ' > ' .$title;
-        return CategoryController::getParentsTree($parent,$title);
-    }
+    
 
     /**
      * Display a listing of the resource.
@@ -31,8 +20,8 @@ class CategoryController extends Controller
     public function index()
     {
         //
-        $data= Category::all();
-        return view('admin.category.index',[
+        $data = Place::all();
+        return view('admin.place.index',[
             'data' => $data
         ]);
     }
@@ -47,7 +36,7 @@ class CategoryController extends Controller
         //
 
         $data= Category::all();
-        return view('admin.category.create',[
+        return view('admin.place.create',[
             'data' => $data
         ]);
     }
@@ -61,30 +50,40 @@ class CategoryController extends Controller
     public function store(Request $request)
     {
         //
-         $data= new Category();
-        $data->parent_id = $request->parent_id;
+         $data= new Place();
+         $data->category_id = $request->category_id;
+         $data->user_id = 0; // $request->category_id;
          $data->title = $request->title;
          $data->keywords = $request->keywords;
          $data->description = $request->description;
          $data->status = $request->status;
+         $data->detail = $request->detail;
+         $data->price = $request->price;
+         $data->quantity = $request->quantity;
+         $data->minquantity = $request->minquantity;
+         $data->tax = $request->tax;
+         $data->city = $request->city;
+         $data->country = $request->country;
+         $data->location = $request->location;
+         $data->slug = $request->slug;
          if ($request->file('image')){
             $data->image = $request->file('image')->store('images');
         }
          $data->save();
-         return redirect('admin/category');
+         return redirect('admin/place');
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Place  $place
      * @return \Illuminate\Http\Response
      */
-    public function show(Category $category,$id)
+    public function show(Place $place,$id)
     {
         //
-        $data= Category::find($id);
-        return view('admin.category.show',[
+        $data= Place::find($id);
+        return view('admin.place.show',[
             'data' => $data
         ]);
     }
@@ -92,15 +91,15 @@ class CategoryController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Place  $place
      * @return \Illuminate\Http\Response
      */
-    public function edit(Category $category,$id)
+    public function edit(Place $place,$id)
     {
         //
-        $data= Category::find($id);
+        $data= Place::find($id);
         $datalist= Category::all();
-        return view('admin.category.edit',[
+        return view('admin.place.edit',[
             'data' => $data,
             'datalist' => $datalist
         ]);
@@ -110,37 +109,54 @@ class CategoryController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Place  $place
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Category $category,$id)
+    public function update(Request $request, Place $place,$id)
     {
         //
-         $data= Category::find($id);
-        $data->parent_id = $request->parent_id;
+         $data= Place::find($id);
+         $data->category_id = $request->category_id;
+         $data->user_id = 0; // $request->category_id;
          $data->title = $request->title;
          $data->keywords = $request->keywords;
          $data->description = $request->description;
          $data->status = $request->status;
+         $data->detail = $request->detail;
+         $data->price = $request->price;
+         $data->quantity = $request->quantity;
+         $data->minquantity = $request->minquantity;
+         $data->tax = $request->tax;
+         $data->city = $request->city;
+         $data->country = $request->country;
+         $data->location = $request->location;
+         
          if ($request->file('image')){
             $data->image = $request->file('image')->store('images');
         }
          $data->save();
-         return redirect('admin/category');
-    }
+         return redirect('admin/place');
+        }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Category  $category
+     * @param  \App\Models\Place  $place
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Category $category,$id)
+    public function destroy(Place $place,$id)
     {
         //
-        $data= Category::find($id);
-        Storage::delete($data->image);
+    //     $data = Place::find($id);
+    //    /*  Storage::delete($data->image);  tek bunu bırak yorum satırında olmazsa*/
+    //     $data->delete();
+    //     return redirect('admin/place');
+    // -------------------------------
+        $data = Place::find($id);
+        if ($data->image && Storage::disk('public')->exists($data->image)) {
+            Storage::delete($data->image);
+        }
         $data->delete();
-        return redirect('admin/category');
+        return redirect('admin/place');
     }
 }
